updatet

+++ WEB/web-races/webapp/src/main/java/by/training/java/grodno/az/webapp/page/ratelinepage/RateLinePage.java
@@ -1,6 +1,8 @@
 package by.training.java.grodno.az.webapp.page.ratelinepage;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import javax.inject.Inject;
 
@@ -10,13 +12,11 @@ import org.apache.wicket.markup.html.link.Link;
 import org.apache.wicket.markup.html.list.ListItem;
 import org.apache.wicket.markup.html.list.ListView;
 
-import by.training.java.grodno.az.data.model.Jockey;
 import by.training.java.grodno.az.data.model.RateLine;
-import by.training.java.grodno.az.service.JockeyService;
+import by.training.java.grodno.az.service.CoefficientService;
 import by.training.java.grodno.az.service.RateLineService;
 import by.training.java.grodno.az.webapp.page.abstractpage.AbstractPage;
 import by.training.java.grodno.az.webapp.page.admin.coefficientpage.CoefficientEditPage;
-import by.training.java.grodno.az.webapp.page.admin.coefficientpage.CoefficientsPage;
 
 @AuthorizeInstantiation(value = { "admin" })
 public class RateLinePage extends AbstractPage {
@@ -25,6 +25,9 @@ public class RateLinePage extends AbstractPage {
 
 	@Inject
 	private RateLineService rateLineService;
+
+	@Inject
+	private CoefficientService coefficientService;
 
 	@Override
 	protected void onInitialize() {
@@ -51,8 +54,14 @@ public class RateLinePage extends AbstractPage {
 
 					@Override
 					public void onClick() {
-						rateLineService.delete(rateLine);
-						setResponsePage(RateLinePage.class);
+						if (canBeDelete(rateLine)) {
+							rateLineService.delete(rateLine);
+							setResponsePage(RateLinePage.class);
+						} else {
+							RateLinePage responsePage = new RateLinePage();
+							responsePage.error(getString("page.rateLinePage.not.delete"));
+							setResponsePage(responsePage);
+						}
 					}
 				});
 
@@ -65,12 +74,23 @@ public class RateLinePage extends AbstractPage {
 				if (rateLines.size() < CoefficientEditPage.MAXQUANTITY) {
 					setResponsePage(new RateLineEditPage());
 				} else {
-					AbstractPage responsePage=new RateLinePage();
+					AbstractPage responsePage = new RateLinePage();
 					warn(getString("all.tableRecords.limit"));
 					setResponsePage(responsePage);
 				}
 			}
 		});
+	}
+
+	private boolean canBeDelete(RateLine rateLine) {
+		boolean result = false;
+		Map<String, Object> findingParemeters = new HashMap<>();
+		findingParemeters.put("rateLineId", rateLine.getId());
+		int count = coefficientService.getCount(findingParemeters);
+		if (count == 0) {
+			result = true;
+		}
+		return result;
 	}
 
 }
